Phase 2: move simulator.py/mps.py into dense_evolution.backends/ (statevector.py + mps.py)
- git mv dense_evolution/simulator.py -> dense_evolution/backends/statevector.py
  (renamed, per the target tree's distinction between backends/ [the compute
  engines] and solvers/ [the algorithms built on top, moving later])
- git mv dense_evolution/mps.py -> dense_evolution/backends/mps.py
- new backends/__init__.py re-exports DenseSVSimulator, MPSSimulator
- backward-compat shims at both old top-level paths (dense_evolution.simulator,
  dense_evolution.mps) -- the rename means the dotted path genuinely changes,
  same as physics/, unlike interop/
- fixed relative imports broken by the move: both files' .circuits.<mod>
  imports (gates/compiler/registry) -> ..circuits.<mod>, one level deeper now
- updated dense_evolution/__init__.py and interop/qiskit_pennylane.py's
  ..simulator import -> ..backends.statevector
- dense_evolution/chunk.py needed no change: its flexible try/except import
  chain already goes through the flat dense_evolution.simulator/.compiler/
  .gates absolute paths, exactly what the shims preserve

Fixed two real regressions found during verification, same private-helper-
import pattern as interop/ and circuits/:
- tests/test_mps.py imported private helpers (_jsd_vectors,
  _vectorized_chi_search, _expand_nonlocal_2q_positions) from the
  dense_evolution.mps shim -- pointed at dense_evolution.backends.mps directly
- tests/test_simulator.py did `import dense_evolution.simulator as sim_mod`
  then monkeypatched HAS_JAX on it -- the real check lives in
  statevector.py's own module globals, so pointed at
  dense_evolution.backends.statevector directly

Full suite: 799 passed, 17 skipped, 1 failed on first run -- MemoryPressureError
(known resource-contention pattern), confirmed via isolated re-run: passes alone.

Part of #76.
