+++ dense_evolution/backends/__init__.py
@@ -1,0 +1,5 @@
+"""Backends subpackage: the dense statevector and MPS compute engines."""
+from .statevector import DenseSVSimulator
+from .mps import MPSSimulator
+
+__all__ = ["DenseSVSimulator", "MPSSimulator"]
